Project restructuring (fixes macOS builds) (#198)
- Create `Include/` folder for all main projects
- Fix some issues here and there (found while moving things)
- All includes are now in `Tactility/` subfolder and must be included with that prefix. This fixes issues with clashing POSIX headers (e.g. `<semaphore.h>` versus Tactility's `Semaphore.h`)

+++ Tactility/Source/service/gui/Gui.cpp
@@ -1,10 +1,11 @@
-#include "Tactility.h"
-#include "service/gui/Gui_i.h"
-#include "service/loader/Loader_i.h"
-#include "lvgl/LvglSync.h"
-#include "RtosCompat.h"
-#include "lvgl/Style.h"
-#include "lvgl/Statusbar.h"
+#include "Tactility/service/gui/Gui_i.h"
+#include "Tactility/service/loader/Loader_i.h"
+#include "Tactility/lvgl/LvglSync.h"
+#include "Tactility/lvgl/Style.h"
+#include "Tactility/lvgl/Statusbar.h"
+
+#include <Tactility/Tactility.h>
+#include <Tactility/RtosCompat.h>
 
 namespace tt::service::gui {
 
@@ -120,7 +121,7 @@ static int32_t guiMain(TT_UNUSED void* p) {
     Gui* local_gui = gui;
 
     while (true) {
-        uint32_t flags = Thread::awaitFlags(GUI_THREAD_FLAG_ALL, EventFlag::WaitAny, portMAX_DELAY);
+        uint32_t flags = Thread::awaitFlags(GUI_THREAD_FLAG_ALL, EventFlag::WaitAny, (uint32_t)portMAX_DELAY);
 
         // Process and dispatch draw call
         if (flags & GUI_THREAD_FLAG_DRAW) {
